Moves `LocalFeedItem` to its own file
